pyproject.toml
[project]
name = "llmops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.129.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.10",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.8",
    "langsmith>=0.7.1",
    "openai>=2.18.0",
    "openevals>=0.1.3",
    "pandas>=3.0.0",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.2.2",
    "structlog>=25.5.0",
]
